requirements.txt
python-dotenv
diffusers
torch
transformers
accelerate
protobuf
sentencepiece
bitsandbytes